Move inspector from primary toolbox to secondary toolbox
