+++ src/ui/components/Toolbox.css
@@ -23,25 +23,10 @@
   border-right: 1px solid var(--theme-splitter-color);
 }
 
-#toolbox-contents {
-  height: 100%;
-  position: relative;
-  overflow: hidden;
-  display: flex;
-  flex-grow: 1;
-  flex-direction: column;
-}
-
 /* Split Console styles */
 .toolbox-top-panels {
   width: 100%;
   height: 100%;
-}
-.toolbox-top-panels .toolbox-panel.active {
-  display: block;
-}
-.toolbox-top-panels .toolbox-panel {
-  display: none;
 }
 
 /* Toolbar Styles */
