+++ src/ui/components/Toolbox.js
@@ -1,14 +1,8 @@
 import React from "react";
-import classnames from "classnames";
 import KeyShortcuts from "devtools/client/shared/key-shortcuts";
-import SplitBox from "devtools/client/shared/components/splitter/SplitBox";
 
 import DebuggerApp from "devtools/client/debugger/src/components/App";
-import InspectorApp from "devtools/client/inspector/components/App";
-import WebConsoleApp from "devtools/client/webconsole/components/App";
 
-import Timeline from "./Timeline";
-import NodePicker from "./NodePicker";
 import { connect } from "react-redux";
 import { actions } from "../actions";
 import { selectors } from "../reducers";
@@ -23,14 +17,10 @@ class Toolbox extends React.Component {
   };
 
   async componentDidMount() {
-    const { selectedPanel } = this.props;
+    const selectedPanel = "debugger";
     await gToolbox.init(selectedPanel);
 
     shortcuts.on("Esc", this.onEscape);
-  }
-
-  get currentTool() {
-    return this.props.selectedPanel;
   }
 
   onEscape = e => {
@@ -45,129 +35,31 @@ class Toolbox extends React.Component {
     this.props.setSplitConsole(open);
   }
 
-  renderInspector() {
-    const { initializedPanels } = this.props;
-    const inspector = gToolbox.getPanel("inspector");
-    let markupView, rulesPanel, layoutPanel, computedPanel;
-
-    if (inspector && initializedPanels.includes("inspector")) {
-      markupView = inspector._inspector.getPanel("markupview").provider;
-      rulesPanel = {
-        id: "ruleview",
-        title: "Rules",
-        panel: inspector._inspector.getPanel("ruleview").provider,
-      };
-      layoutPanel = {
-        id: "layoutview",
-        title: "Layout",
-        panel: inspector._inspector.getPanel("layoutview").provider,
-      };
-      computedPanel = {
-        id: "computedview",
-        title: "Computed",
-        panel: inspector._inspector.getPanel("computedview").provider,
-      };
-    }
-    return (
-      <>
-        <InspectorApp
-          markupView={markupView}
-          rulesPanel={rulesPanel}
-          layoutPanel={layoutPanel}
-          computedPanel={computedPanel}
-        />
-        <div id="tabpanels" style={{ display: "none" }}>
-          <div id="sidebar-panel-computedview" className="theme-sidebar inspector-tabpanel">
-            <div id="computed-toolbar" className="devtools-toolbar devtools-input-toolbar">
-              <div className="devtools-searchbox">
-                <input
-                  id="computed-searchbox"
-                  className="devtools-filterinput"
-                  type="search"
-                  data-localization="placeholder=inspector.filterStyles.placeholder"
-                />
-                <button
-                  id="computed-searchinput-clear"
-                  className="devtools-searchinput-clear"
-                ></button>
-              </div>
-              <div className="devtools-separator"></div>
-              <input id="browser-style-checkbox" type="checkbox" className="includebrowserstyles" />
-              <label
-                id="browser-style-checkbox-label"
-                htmlFor="browser-style-checkbox"
-                data-localization="content=inspector.browserStyles.label"
-              ></label>
-            </div>
-
-            <div id="computed-container">
-              <div id="computed-container-focusable" tabIndex="-1">
-                <div
-                  id="computed-property-container"
-                  className="devtools-monospace"
-                  tabIndex="0"
-                  dir="ltr"
-                ></div>
-                <div
-                  id="computed-no-results"
-                  className="devtools-sidepanel-no-result"
-                  hidden=""
-                  data-localization="content=inspector.noProperties"
-                ></div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </>
-    );
-  }
-
   renderToolbar() {
-    const { selectedPanel } = this.props;
-
     return (
       <div id="toolbox-toolbar">
-        {/* <NodePicker /> */}
         <div
-          className={classnames("toolbar-panel-button", {
-            active: selectedPanel == "inspector",
-          })}
-          id="toolbox-toolbar-inspector"
-          onClick={() => gToolbox.selectTool("inspector")}
-        >
-          <div className="toolbar-panel-icon"></div>
-        </div>
-        <div
-          className={classnames("toolbar-panel-button", {
-            active: selectedPanel == "debugger",
-          })}
+          className={"toolbar-panel-button active"}
           id="toolbox-toolbar-debugger"
           onClick={() => gToolbox.selectTool("debugger")}
         >
           <div className="toolbar-panel-icon"></div>
         </div>
-        {selectedPanel == "debugger" ? (
-          <>
-            <div
-              className="toolbar-panel-button toolbar-panel-button-sub"
-              onClick={() => this.setState({ debuggerMode: "explorer" })}
-            >
-              <div
-                className="img document toolbar-panel-icon"
-                style={{ background: "#c9c9cb", height: "20px", width: "20px" }}
-              />
-            </div>
-            <div
-              className="toolbar-panel-button toolbar-panel-button-sub"
-              onClick={() => this.setState({ debuggerMode: "debug" })}
-            >
-              <div
-                style={{ height: "20px", width: "20px" }}
-                className="img log toolbar-panel-icon"
-              />
-            </div>
-          </>
-        ) : null}
+        <div
+          className="toolbar-panel-button toolbar-panel-button-sub"
+          onClick={() => this.setState({ debuggerMode: "explorer" })}
+        >
+          <div
+            className="img document toolbar-panel-icon"
+            style={{ background: "#c9c9cb", height: "20px", width: "20px" }}
+          />
+        </div>
+        <div
+          className="toolbar-panel-button toolbar-panel-button-sub"
+          onClick={() => this.setState({ debuggerMode: "debug" })}
+        >
+          <div style={{ height: "20px", width: "20px" }} className="img log toolbar-panel-icon" />
+        </div>
       </div>
     );
   }
@@ -202,43 +94,20 @@ class Toolbox extends React.Component {
   }
 
   render() {
-    const { selectedPanel, splitConsoleOpen } = this.props;
     const { debuggerMode } = this.state;
 
     const topPanels = (
       <div className="toolbox-top-panels">
-        <div
-          className={classnames("toolbox-panel", {
-            active: selectedPanel == "debugger",
-          })}
-          id="toolbox-content-debugger"
-        >
+        <div className="toolbox-panel" id="toolbox-content-debugger">
           <DebuggerApp debuggerMode={debuggerMode} />
-        </div>
-        <div
-          className={classnames("toolbox-panel theme-body", {
-            active: selectedPanel == "inspector",
-          })}
-          id="toolbox-content-inspector"
-        >
-          {this.renderInspector()}
         </div>
       </div>
     );
 
     return (
-      <div id="toolbox" className={`${selectedPanel}`}>
+      <div id="toolbox">
         {this.renderToolbar()}
-        {this.props.toolboxExpanded && (
-          <div
-            id="toolbox-contents"
-            className={classnames("", {
-              splitConsole: selectedPanel != "console" && splitConsoleOpen,
-            })}
-          >
-            {topPanels}
-          </div>
-        )}
+        {topPanels}
       </div>
     );
   }
@@ -246,12 +115,9 @@ class Toolbox extends React.Component {
 export default connect(
   state => ({
     initializedPanels: selectors.getInitializedPanels(state),
-    selectedPanel: selectors.getSelectedPanel(state),
-    splitConsoleOpen: selectors.isSplitConsoleOpen(state),
     toolboxExpanded: selectors.getToolboxExpanded(state),
   }),
   {
-    setSplitConsole: actions.setSplitConsole,
     setSelectedPanel: actions.setSelectedPanel,
   }
 )(Toolbox);
